requirements.txt
accelerate==1.0.1
aiohttp~=3.10.2
datasets==3.2.0
huggingface-hub==0.27.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
omegaconf==2.3.0
pandas>=2.2.2
pillow==10.4.0
pyarrow>=15.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
spacy==3.7.4
sentencepiece==0.2.0
sentence_transformers==3.3.1
tokenizers==0.21.0
tqdm==4.66.4
transformers==4.48.3
typing_extensions==4.12.2
wandb>=0.16.6
peft==0.12.0
dataclasses~=0.6
numpy~=1.26.4
nltk~=3.9.1
requests~=2.32.3
iopath~=0.1.10